pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "flask-wtf>=1.2.2",
    "wtforms>=3.2.1",
    "requests>=2.32.3",
    "feedparser>=6.0.11",
    "inference-sdk>=0.47.0",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.40",
    "werkzeug>=3.1.3",
    "opencv-python-headless>=4.11.0.86",
    "sendgrid>=6.11.0",
    "reportlab>=4.4.0",
    "pandas>=2.2.3",
]
